requirements.txt
python-telegram-bot>=20.0
google-genai>=0.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0
